+++ src/osu/data/gosu/gosumemoryclient.h
